Restructured entire project setup
1. Created folders for each file type:
    assets
    html
    javascripts
    styles

2. Updated paths

3. Updated html structures to be standardized

4. Updated CSS structures to match original style

5. Updated script.js to fetch data for header and footer to display on pages

6. Created footer.html and header.html which will have the static information such as navbar on every page
