+++ src/Assets/footer.html
@@ -1,0 +1,4 @@
+<!--Footer Section-->
+<footer id="footer">
+    <p>&copy; 2024 SnackTrack. All rights reserved.</p>
+</footer>
