+++ src/Assets/header.html
@@ -1,0 +1,10 @@
+<div class="header">
+    <h1 class="title">SnackTrack</h1>
+    <div class="navbar">
+        <a href="home.html" class="navbar_element">Home</a>
+        <a href="about.html" class="navbar_element">About Us</a>
+        <a href="help.html" class="navbar_element">Help</a>
+        <a href="login.html" class="navbar_element">Login</a>
+        <a href="join.html" class="navbar_element">Join Now</a>
+    </div>
+</div>
