+++ src/Javascripts/script.js
@@ -1,0 +1,16 @@
+// Load header and navigation buttons
+fetch('../Assets/header.html')
+    .then(response => response.text())
+    .then(data => {
+        // Insert the fetched header HTML into the header section
+        document.getElementById('header').innerHTML = data;
+    })
+    .catch(error => console.error('Error loading header:', error)); // Error handling
+
+// Load footer
+fetch('../Assets/footer.html')
+    .then(response => response.text())
+    .then(data => {
+        document.getElementById('footer').innerHTML = data;
+    })
+    .catch(error => console.error('Error loading footer:', error)); // Error handling
