--- src/about.html
@@ -1,36 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Home</title>
-    <link rel="stylesheet" href="styles.css">
-</head>
-<body>
-    <div class="header">
-    <h1 class="title">SnackTrack</h1>
-        <div class="navbar"></div>
-            <a href="/index.html" class="navbar_element">Home</a>
-            <a href="/about.html" class="navbar_element">About Us</a>
-            <a href="/help.html" class="navbar_element">Help</a>
-            <a href="/login.html" class="navbar_element">Login</a>
-            <a href="/join.html" class="navbar_element">Join Now</a>
-        </div>
-    </div>
-    <p>
-        SnackTrack will help users monitor their fitness journeys by 
-        providing an accessible platform to log their food, exercise, 
-        and overall health metrics, allowing them to achieve their fitness 
-        goals. Our proposed system will allow users to input personal 
-        details such as their daily caloric intake and exercise routines, 
-        and track them over time. The app will provide an easy-to-navigate 
-        interface where users can log their meals and workouts, receive 
-        feedback, and view progress through attractive visualizations. 
-        We will use the Sport and Fitness API to access reliable nutrition 
-        and fitness data for food items and exercises. This will simplify 
-        the process for users by providing a comprehensive database for 
-        tracking and storing their data.
-
-    </p>
-</body>
-</html>
