+++ src/html/about.html
@@ -1,0 +1,47 @@
+<!DOCTYPE html>
+<html lang="en">
+    <head>
+        <meta charset="UTF-8">
+        <meta name="viewport" content="width=device-width, initial-scale=1.0">
+        <title>About | SnackTrack</title>
+        <link rel="stylesheet" href="../styles/styles.css">
+        <script defer src="../Javascripts/script.js"></script>
+    </head>
+    <body>
+        <!-- Header Section -->
+        <header id="header"></header>
+
+        <!-- Main Content Section -->
+        <div class="main-container">
+            <section>
+                <h2>Welcome to SnackTrack!</h2>
+                <article>
+                    <p>
+                        SnackTrack will help users monitor their fitness journeys by 
+                        providing an accessible platform to log their food, exercise, 
+                        and overall health metrics, allowing them to achieve their fitness 
+                        goals. Our proposed system will allow users to input personal 
+                        details such as their daily caloric intake and exercise routines, 
+                        and track them over time. 
+                    </p>
+                    <p>
+                        The app will provide an easy-to-navigate interface where users 
+                        can log their meals and workouts, receive feedback, and view 
+                        progress through attractive visualizations. 
+                    </p>
+                    <p>
+                        We will use the Sport and Fitness API to access reliable nutrition 
+                        and fitness data for food items and exercises. This will simplify 
+                        the process for users by providing a comprehensive database for 
+                        tracking and storing their data.
+                    </p>
+                </article>
+            </section>
+        
+        </div>
+        
+
+        <!-- Footer Section -->
+        <footer id="footer"></footer>
+    </body>
+</html>
