+++ src/html/help.html
@@ -1,0 +1,24 @@
+<!DOCTYPE html>
+<html lang="en">
+    <head>
+        <meta charset="UTF-8">
+        <meta name="viewport" content="width=device-width, initial-scale=1.0">
+        <title>Help | SnackTrack</title>
+        <link rel="stylesheet" href="../styles/styles.css">
+        <script defer src="../Javascripts/script.js"></script>
+    </head>
+    <body>
+        <!-- Header Section -->
+        <header id="header"></header>
+        
+        <!-- Main Content Section -->
+        <div class="main-container">
+
+
+
+        </div>
+        
+        <!-- Footer Section -->
+        <footer id="footer"></footer>
+    </body>
+</html>
