+++ src/html/home.html
@@ -1,0 +1,30 @@
+<!DOCTYPE html>
+<html lang="en">
+    <head>
+        <meta charset="UTF-8">
+        <meta name="viewport" content="width=device-width, initial-scale=1.0">
+        <title>Home | SnackTrack</title>
+        <link rel="stylesheet" href="../styles/styles.css">
+        <script defer src="../Javascripts/script.js"></script>
+    </head>
+    <body>
+        <!-- Header Section -->
+        <header id="header"></header>
+        
+        
+        <!-- Main Content Section -->
+        <main class="main-container">
+            <h2>Track Your Snacks and Stay on Track!</h2>
+            <p>
+                Welcome to SnackTrack! Our platform helps you monitor your fitness journey by providing an accessible way to log your food, exercise, and overall health metrics.
+            </p>
+            <p>
+                Visit our About Us page to learn more about our mission! If you're ready to experience a healthier lifestyle in an organized and educational way, create an account with us now!
+            </p>
+        </main>
+        
+
+        <!-- Footer Section -->
+        <footer id="footer"></footer>
+    </body>
+</html>
