--- src/index.html
@@ -1,27 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Home</title>
-    <link rel="stylesheet" href="styles.css">
-</head>
-<body>
-    <div class="header">
-    <h1 class="title">SnackTrack</h1>
-        <div class="navbar"></div>
-            <a href="/index.html" class="navbar_element">Home</a>
-            <a href="/about.html" class="navbar_element">About Us</a>
-            <a href="/help.html" class="navbar_element">Help</a>
-            <a href="/login.html" class="navbar_element">Login</a>
-            <a href="/join.html" class="navbar_element">Join Now</a>
-        </div>
-    </div>
-    <h2>Track your snacks and stay on track!</h2>
-    <p>
-        Visit our About Us page to learn more about our mission!
-        If you want to experience a healthier lifestyle in an 
-        organized and educational way, create an account with us now!
-    </p>
-</body>
-</html>
