--- src/login.html
@@ -1,21 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Home</title>
-    <link rel="stylesheet" href="styles.css">
-</head>
-<body>
-    <div class="header">
-    <h1 class="title">SnackTrack</h1>
-        <div class="navbar"></div>
-            <a href="/index.html" class="navbar_element">Home</a>
-            <a href="/about.html" class="navbar_element">About Us</a>
-            <a href="/help.html" class="navbar_element">Help</a>
-            <a href="/login.html" class="navbar_element">Login</a>
-            <a href="/join.html" class="navbar_element">Join Now</a>
-        </div>
-    </div>
-</body>
-</html>
