--- src/styles.css
@@ -1,40 +1,0 @@
-body {
-    background: #a0a0a0;
-    font-family: Verdana;
-}
-
-a {
-    color: white;
-    text-decoration: none;
-    padding: 10px;
-}
-
-p, h2 {
-    color: white;
-}
-
-.header {
-    display: flex; 
-    justify-content: space-between; 
-    align-items: center; 
-    padding: 10px 50px;
-    background-color: #5e5e5e;
-    color: white;
-}
-
-.title {
-    float: left;
-}
-
-.navbar_element {
-    float: right;
-}
-
-
-.navbar_element:hover {
-    color: #cecece;
-    border: solid;
-    border-color: white;
-    border-radius: 5px;
-    padding: 8px;
-}
